添加了图片上传的验证，也修改了一点工具类的代码 完善了导入 excel 的方法，既当有相同登录名的用户导入时，让其导入失败，并且在导入 excel 方法的最后执行了删除服务器上 excel 文件的操作 在 index.jsp 上删除了 2 行 css 的代码， 因为该资源失效了， 导致 index.jsp 无法请求访问

+++ src/main/java/com/work/util/FileUploadUtil.java
@@ -49,14 +49,21 @@ public class FileUploadUtil {
      * @param uploadFile
      * @param
      */
-    public static void uploadProductPicture(MultipartFile uploadFile,Goods goods){
+    public static boolean uploadProductPicture(MultipartFile uploadFile,Goods goods){
+        String fileName = null;
         if (!uploadFile.isEmpty()){
+            fileName = uploadFile(uploadFile, PRODUCT_PATH);
+            String suffix = fileName.substring(36, 40);
+            if(!".jpg".equals(suffix) && !".png".equals(suffix) && !".bpn".equals(suffix)){
+                System.out.println("不通过  ------------------------------");
+                return false;
+            }
             if (goods.getPicture()!=null){
                 deleteFile(PRODUCT_PATH,goods.getPicture());
             }
-            goods.setPicture(uploadFile(uploadFile, PRODUCT_PATH));
-
+            goods.setPicture(fileName);
         }
+        return true;
     }
 	public static String uploadFile(MultipartFile uploadFile, String savePath) {
         RequestAttributes ar = RequestContextHolder.getRequestAttributes();
